Added some error checking to make sure the program exits if anything goes wrong
git-svn-id: svn+ssh://cvs1/svn/listen/trunk@334 4692729f-32a0-46b7-9832-1bf53c4881b5

+++ release/deployFromHudson.sh
@@ -31,19 +31,45 @@ else
 fi
 
 # Get upstream artifact
-wget -nv ${UPSTREAM_URL}/artifact//*zip*/listen.zip
-unzip listen.zip
-export listenRPM=`find ./archive -name "listen-*.rpm"`
+wget -nv ${UPSTREAM_URL}/artifact//*zip*/listen.zip && unzip listen.zip
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
 
 # Get last successful uia
-wget -nv ${UIA_URL}/artifact//*zip*/uia.zip
-unzip uia.zip
+wget -nv ${UIA_URL}/artifact//*zip*/uia.zip && unzip uia.zip
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+
+export listenRPM=`find ./archive -name "listen-*.rpm"`
 export UIARPM=`find ./archive -name "uia*.rpm"`
 
 ssh ${SSH_OPTS} root@${listenserver} "rm -f /interact/packages/listen*rpm"
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+
 scp ${listenRPM} ${UIARPM} ${listenserver}:/interact/
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+
 scp deployListen.py ${listenserver}:/root/
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+
 
 # Invoke deploy -- do all phases
 ssh ${SSH_OPTS} root@${listenserver} "/root/deployListen.py --listenserver=${listenserver} --insaserver=${insaserver} --phase=all"
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
